Move isKeyModified from constants to statics
